pyproject.toml
[tool.poetry]
name = "fwh-orders"
version = "2.0.0"
description = "It's not always good to start from scratch .. lets have some sane defaults .."
authors = ["Andy Grabow <andy@freilandkiwis.de>"]
license = "MIT"
readme = 'README.md'
package-mode = false

repository = "https://github.com/kakulukia/fwh-orders"
homepage = "https://github.com/kakulukia/fwh-orders"

[tool.poetry.dependencies]
django = "^5.2.6"
django-axes = "^8.0.0"
django-compressor = "^4.5.1"
django-extensions = "^4.1"
django-kronos = "^1.0"
django-post-office = "^3.10.1"
django-secrets = "^2.0.1"
django-tasks = "^0.8.1"
djangorestframework = "^3.16.1"
huepy = "^1.2.1"
pendulum = "^3.1.0"
Pillow = "^11.3.0"
pypugjs = "^5.12.0"
python = "^3.11"
requests = "^2.32.5"
sentry-sdk = "^2.37.1"
whitenoise = "^6.10.0"
django-loginas = "^0.3.12"
# django-model-utils - use this if a status field is needed and maybe also a status_changed field
gunicorn = "^23.0.0"
django-allauth = "^65.13.0"
django-environ = "^0.12.0"
reportlab = "^4.4.4"
openpyxl = "^3.1.5"
django-crontab = "^0.7.1"
pytz = "^2025.2"

[tool.poetry.group.dev.dependencies]
django-browser-reload = "^1.19.0"
django-debug-toolbar = "^6.0.0"
icecream = "^2.1.7"
ipdb = "^0.13.13"
pre-commit = "^4.3.0"
ruff = "^0.13.0"
fab-classic = "^1.20.1"  # poetry add --dev "git+https://github.com/kakulukia/fab-classic.git#master" (ssh key fwding)
toml = "^0.10.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
exclude = [
    "migrations",
    "node_modules",
    ".venv",
]

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001"
]
